requirements.txt
discord.py==1.3.3
peewee==3.11.2
